move this logic to the template
each component needs different inputs and this was the clearest way I
could come up with to make sure the args were explicit enough to
understand

this uses the progress service, which is injected into the audio survey
manager (and potentially other "survey manager" components), to pass
certain args, in addition any args provided by the route.

also the `next` action is passed in, whcih is controlled by the survey
manager but provided args by the components themselves.