adds fastboot process for production
this will handle http traffic to the app in production. When it boots,
the process will download a zip of the built app from the specified S3
bucket, unzip it, and start a pool of workers, each running separate
instances of the app to handled incoming requests.

the configired S3Notifier periodically polls the S3 bucket to check for
a new version of the built assets, and if it finds a different version,
will notify the downloader to grab it and reload the running instances.